refactor: change the querydsl dto package according to the agreed-upon rules
