+++ petlog-api/src/main/java/com/ppp/api/user/dto/response/UserResponse.java
@@ -2,7 +2,7 @@ package com.ppp.api.user.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.ppp.domain.user.User;
-import com.ppp.domain.user.UserDao;
+import com.ppp.domain.user.dto.UserDto;
 import com.ppp.domain.user.UserDocument;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
@@ -39,7 +39,7 @@ public record UserResponse(
                 .build();
     }
 
-    public static UserResponse from(UserDao user, String currentUserId) {
+    public static UserResponse from(UserDto user, String currentUserId) {
         return UserResponse.builder()
                 .id(user.getId())
                 .nickname(user.getNickname())
